requirements-test.txt
pytest
pyyaml
requests
biopython
numpy
pandas
plotly
